requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python-headless
requests
yt-dlp
python-multipart
exifread
